feat/refector: adição de tratativas de erros, login e loggoff, relacionamento básico de fkś

+++ backend/projectfindr/src/main/java/findr/projectfindr/controller/CompanyController.java
@@ -2,6 +2,7 @@ package findr.projectfindr.controller;
 
 import findr.projectfindr.model.Contactor;
 import findr.projectfindr.repository.ContactorRepository;
+import findr.projectfindr.repository.ProjectRepository;
 import findr.projectfindr.resposta.LoginResposta;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -13,32 +14,59 @@ public class CompanyController {
     @Autowired
     private ContactorRepository bd;
 
+    @Autowired
+    private ProjectRepository projects;
+
     @PostMapping
     public ResponseEntity addUserContactorCompany(@RequestBody Contactor company) {
-        bd.save(company);
-        return ResponseEntity.status(200).build();
+        try {
+            bd.save(company);
+        }catch (Exception e){
+            return ResponseEntity.status(406).build();
+        }
+        return ResponseEntity.status(201).build();
     }
 
     @GetMapping //verificar responseEntity
     public ResponseEntity getCompany() {
+        if (bd.findAll().isEmpty()){
+            return ResponseEntity.status(204).body(bd.findAll());
+        }
         return ResponseEntity.status(200).body(bd.findAll());
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity deleteUserContactorCompany(@PathVariable Long id) {
-        bd.deleteById(id);
-        return ResponseEntity.status(200).build();
+        if (bd.existsById(id)){
+            bd.deleteById(id);
+            return ResponseEntity.status(200).body(bd.findAll());
+        }
+        return ResponseEntity.status(204).build();
     }
 
     @PostMapping("/login")
     public ResponseEntity setLoginCompany(@RequestBody LoginResposta login) {
-        bd.findByEmailAndPassword(login.getEmail(), login.getPassword());
-        return ResponseEntity.status(200).build();
+        if (!bd.findByEmailAndPassword(login.getEmail(), login.getPassword()).isEmpty()){
+            bd.atualizarOnline(login.getEmail(), login.getPassword(), true);
+            return ResponseEntity.status(201).build();
+        }
+        return ResponseEntity.status(204).build();
     }
 
     @PostMapping("/logoff")
     public ResponseEntity setLogoffCompany(@RequestBody LoginResposta logoff) {
-        bd.findByEmailAndPassword(logoff.getEmail(), logoff.getPassword());
-        return ResponseEntity.status(200).build();
+        if (!bd.findByEmailAndPassword(logoff.getEmail(), logoff.getPassword()).isEmpty()){
+            bd.atualizarOnline(logoff.getEmail(), logoff.getPassword(), false);
+            return ResponseEntity.status(201).build();
+        }
+        return ResponseEntity.status(204).build();
     }
+
+   @GetMapping("/projects/{idCompany}")
+   public ResponseEntity myProjects(@PathVariable Integer idCompany){
+        if (projects.findByFkContactor(idCompany).isEmpty()){
+            return ResponseEntity.status(204).build();
+        }
+        return ResponseEntity.status(200).body(projects.findByFkContactor(idCompany));
+   }
 }
