+++ backend/projectfindr/src/main/java/findr/projectfindr/controller/ProjectController.java
@@ -16,21 +16,30 @@ public class ProjectController {
     private ProjectRepository bd;
 
     @PostMapping
-    public ResponseEntity addProjectCompany(
-            @RequestBody @Valid ProjectModel project){
+    public ResponseEntity addProjectCompany(@RequestBody @Valid ProjectModel project) {
+        try {
             bd.save(project);
+        }catch (Exception e){
+            return ResponseEntity.status(406).build();
+        }
         return ResponseEntity.status(200).build();
     }
 
     @GetMapping
-    public ResponseEntity getProject(){
+    public ResponseEntity getProject() {
+        if (bd.findAll().isEmpty()){
+            return ResponseEntity.status(204).body(bd.findAll());
+        }
         return ResponseEntity.status(200).body(bd.findAll());
     }
 
     @DeleteMapping("/{idProject}")
-    public ResponseEntity deletProject(@PathVariable Long idProject){
-        bd.deleteById(idProject);
-        return ResponseEntity.status(200).build();
+    public ResponseEntity deleteProject(@PathVariable Long idProject) {
+        if (bd.existsById(idProject)){
+            bd.deleteById(idProject);
+            return ResponseEntity.status(200).body(bd.findAll());
+        }
+        return ResponseEntity.status(204).build();
     }
 
 }
